src/app/requirements.txt
dash-bootstrap-components
databricks-sql-connector>=3.0
databricks-sdk>=0.30
